lando.workers: move legacy code to new app
