+++ src/lando/api/tests/test_transplants.py
@@ -734,7 +734,7 @@ def test_integrated_transplant_records_approvers_peers_and_owners(
     # Mock a few mots-related things needed by the landing worker.
     # First, mock path existance.
     mock_path = MagicMock()
-    monkeypatch.setattr("lando.api.legacy.workers.landing_worker.Path", mock_path)
+    monkeypatch.setattr("lando.workers.landing_worker.Path", mock_path)
     (mock_path(repo.path) / "mots.yaml").exists.return_value = True
 
     # Then mock the directory/file config.
