+++ src/lando/headless_api/management/commands/start_automation_worker.py
@@ -1,9 +1,9 @@
 import logging
 
-from lando.api.legacy.workers.automation_worker import AutomationWorker
 from lando.main.management.commands.start_worker import (
     Command as StartWorkerCommand,
 )
+from lando.workers.automation_worker import AutomationWorker
 
 logger = logging.getLogger(__name__)
