+++ src/lando/workers/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class WorkersConfig(AppConfig):
+    name = "lando.workers"
